requirements.txt
Django==2.2.10
mysql-connector-python
django-bootstrap4
django-bootstrap-modal-forms
django-bootstrap-datepicker-plus
django-filter==2.4.0
django-betterforms==1.2
xhtml2pdf==0.2.5
